Add files via upload
Objective in this game is to score as many points as possible within 30 seconds (when the game ends automatically). A card will appear somewhere on the board, and if it falls under the description given at the top right, you can click on it to score a point. If you click on a card when it doesn't fit the given description, you get an automatic game over. An ace counts as an odd/numbered card.